[AIRRunner] Fix an LLP64 truncation that broke 26 runner tests on Windows
air-runner's port::set_data_rate took `long`, while its data_rate member is a
double and both of its callers pass one.  On LP64 Linux that narrowing is
lossless -- long is 64 bits -- so the bug was invisible there.  On Windows,
LLP64 makes long 32 bits, and the arch models carry bytes_per_second values of
4e9 and 1e11.  Both overflow int32, and MSVC's out-of-range double->integer
conversion yields INT32_MIN, so every affected link came back with a data rate
of -2147483648 bytes/s.

A negative rate gives a negative transfer cost.  getTransferCost returns
(uint64_t)ceil(seconds * cps), so the negative cycle count wrapped to ~2^64,
simulated time jumped to 18446744073709536365, and the run stalled after one
event instead of reaching the launch terminator.  26 of the 56 Util/Runner
tests failed this way on Windows; dma.mlir emitted 1 of its expected 32
DmaMemcpyNdOp trace events.

Instrumented to confirm rather than inferred: the cost function reported
bps=-2.147484e+09 with vol=16384 dw=2 bytes=32768, i.e. exactly INT32_MIN,
against a JSON value of 1e11.

set_data_rate now takes double, matching the member and both call sites.  No
behaviour change on Linux: every bytes_per_second / bytes_per_cycle in the tree
is 1e9, 4e9 or 1e11 -- integers far inside double's exact range -- so the value
that used to arrive via a 64-bit long arrives identically via double.  This is
the only bare `long` left in the runner.

Also start running tests in the Windows CI job.  It built and ran nothing,
which is the whole reason this survived: a Windows-only miscompile of the cost
model could not turn any job red.  check-air-mlir is the suite that covers it
(air-opt / air-runner + FileCheck, no NPU, no XRT, no Vitis, ~30s).
check-air-python is left out on purpose -- api/errors.py passes but takes ~74s
against the suite's 30s lit timeout, which is a separate problem.

Gate.  On Windows, against origin/main at the pinned toolchain
(mlir_aie 1.4.3.dev55+g10767b5): Util/Runner goes 26 failed / 28 passed ->
0 failed / 54 passed, and check-air-mlir goes 511 passed + 26 failed ->
537 passed + 0 failed, exit 0.

+++ mlir/lib/Util/Runner/Resource.cpp
@@ -106,7 +106,14 @@ public:
 
   void add_connection(port *p) { connected_ports[p->name] = p; }
 
-  void set_data_rate(long bytes_per_cycle) {
+  // double, not an integer type: data_rate is a double, both callers pass one,
+  // and the JSON carries bytes_per_second values in the 1e11 range. Taking
+  // `long` here narrowed every rate through a 32-bit int on Windows -- LLP64
+  // makes long 32 bits, where LP64 Linux makes it 64 -- so 1e11 overflowed and
+  // MSVC's out-of-range conversion produced INT32_MIN. A negative rate gives a
+  // negative transfer cost, which the unsigned cycle counters then wrapped to
+  // ~2^64, stalling the simulation. Linux never saw it.
+  void set_data_rate(double bytes_per_cycle) {
     this->data_rate = bytes_per_cycle;
   }
